test: move v12 tests from v11 and update snapshots
