+++ packages/amplify-migration-tests/src/migration-helpers/check-version.ts
@@ -34,4 +34,4 @@ export const validateVersionsForMigrationTest = async (): Promise<void> => {
  *
  * ext migrate flow: https://github.com/aws-amplify/amplify-cli/pull/8806
  */
-export const allowedVersionsToMigrateFrom = ['8.2.0', '10.5.1', '11.0.0'];
+export const allowedVersionsToMigrateFrom = ['8.2.0', '10.5.1', '12.0.0'];
